# Using Smart pointer to optimizer memory usage of dyGraph