… paddle/fluid/pybind/imperative.cc
… python/paddle/fluid/dygraph/tracer.py
… python/paddle/fluid/tests/unittests/test_imperative_recurrent_usage.py