This PR use smart pointer to hold grads and release the fwd vars. In this way, we can save lots of memory:

Test on [Resnet](https://github.com/PaddlePaddle/models/tree/develop/dygraph/resnet), with Tesla K40m(Mem 11439MiB):

Without smart pointer : max batch size -> 56
With smart pointer: max batch size -> 99

And We test the convergence of several models under `models/dygraph` the result is as follow:

Mnist:
<img width="680" alt="Mnist" src="https://user-images.githubusercontent.com/22361972/58759991-dfd4cd80-8564-11e9-9b28-a8a017860e0f.png">
ptbRNN:
<img width="980" alt="ptb" src="https://user-images.githubusercontent.com/22361972/58759994-e6634500-8564-11e9-9b37-3606412db7cc.png">
reinforcement learning:
<img width="949" alt="rl" src="https://user-images.githubusercontent.com/22361972/58760000-f4b16100-8564-11e9-978f-11f4a5fe31dc.png">
transformer:
<img width="789" alt="transformer" src="https://user-images.githubusercontent.com/22361972/58760007-fe3ac900-8564-11e9-99a7-2e9e6597367b.png">