@JiabinYang 2019-06-03
This Commit will speed up framework by 20%:

The test result is based on Tesla K40m 

[PTB_model](https://github.com/PaddlePaddle/models/tree/develop/dygraph/pbt_lm):
before: 82.2777 s per 1/10 of all data
after: 62.12738 s per 1/10 of  all data

[Resnet](https://github.com/PaddlePaddle/models/tree/develop/dygraph/resnet)Batch size 50
before: 1.49s per batch
after: 1.26s per batch